Loki
In Norse mythology, Loki (Anglicized (/ˈloʊki/)), Loptr, or Hveðrungr
is a god or jötunn (or both). Loki is the son of Fárbauti and Laufey,
and the brother of Helblindi and Býleistr. By the jötunn Angrboða, Loki
is the father of Hel, the wolf Fenrir, and the world serpent
Jörmungandr. By his wife Sigyn, Loki is the father of Narfi and/or
Nari. By the stallion Svaðilfari, Loki is the mother—giving birth in
the form of a mare—to the eight-legged horse Sleipnir. In addition,
Loki is referred to as the father of Váli in Prose Edda, though this
source also refers to Odin as the father of Váli twice, and Váli is
found mentioned as a Son of Loki only once.

Loki's relation with the gods varies by source; Loki sometimes assists
the gods and sometimes behaves in a malicious manner towards them. Loki
is a shape shifter and in separate incidents he appears in the form of
a salmon, a mare, a fly, and possibly an elderly woman named Þökk (Old
Norse 'thanks'). Loki's positive relations with the gods end with his
role in engineering the death of the god Baldr and Loki is eventually
bound by Váli with the entrails of one of his sons. In both the Poetic
Edda and the Prose Edda, the goddess Skaði is responsible for placing a
serpent above him while he is bound. The serpent drips venom from above
him that Sigyn collects into a bowl; however, she must empty the bowl
when it is full, and the venom that drips in the meantime causes Loki
to writhe in pain, thereby causing earthquakes. With the onset of
Ragnarök, Loki is foretold to slip free from his bonds and to fight
against the gods among the forces of the jötnar, at which time he will
encounter the god Heimdallr and the two will slay each other.

Loki is referred to in the Poetic Edda, compiled in the 13th century
from earlier traditional sources; the Prose Edda and Heimskringla,
written in the 13th century by Snorri Sturluson; the Norwegian Rune
Poems, in the poetry of skalds, and in Scandinavian folklore. Loki may
be depicted on the Snaptun Stone, the Kirkby Stephen Stone, and the
Gosforth Cross. Loki's origins and role in Norse mythology, which some
scholars have described as that of a trickster god, have been much
debated by scholars. Loki has been depicted in or is referenced in a
variety of media in modern popular culture.