Changing MM relationships of Sigma Replacements: Navigation patch
The first patch still tries to run even if Sigma Replacements: Navigation is not installed. I'd recommend changing the patch to only run if SR:N is installed but also require the patch to run at the end of the patching order so that it is also applied to any additional mod parts that the user may be using.